+++ src/Http/Requests/UpdateCustomerRequest.php
@@ -23,5 +23,6 @@ class UpdateCustomerRequest extends RevolutRequest
         public ?string $fullName = null,
         public ?string $email = null,
         public ?string $phone = null,
+        public ?string $dateOfBirth = null,
     ) {}
 }